LayoutTests/imported/w3c/web-platform-tests/css/mediaqueries/forced-colors-expected.txt


FAIL Should be parseable in a CSS stylesheet: '(forced-colors)' assert_true: expected true got false
FAIL Should be parseable in a CSS stylesheet: '(forced-colors: none)' assert_true: expected true got false
FAIL Should be parseable in a CSS stylesheet: '(forced-colors: active)' assert_true: expected true got false
PASS Should not be parseable in a CSS stylesheet: '(forced-colors: 0)' 
PASS Should not be parseable in a CSS stylesheet: '(forced-colors: no-preference)' 
PASS Should not be parseable in a CSS stylesheet: '(forced-colors: 10px)' 
PASS Should not be parseable in a CSS stylesheet: '(forced-colors: active 0)' 
PASS Should not be parseable in a CSS stylesheet: '(forced-colors: none active)' 
PASS Should not be parseable in a CSS stylesheet: '(forced-colors: active/none)' 
FAIL Should be parseable in JS: '(forced-colors)' assert_true: expected true got false
FAIL Should be parseable in JS: '(forced-colors: none)' assert_true: expected true got false
FAIL Should be parseable in JS: '(forced-colors: active)' assert_true: expected true got false
PASS Should not be parseable in JS: '(forced-colors: 0)' 
PASS Should not be parseable in JS: '(forced-colors: no-preference)' 
PASS Should not be parseable in JS: '(forced-colors: 10px)' 
PASS Should not be parseable in JS: '(forced-colors: active 0)' 
PASS Should not be parseable in JS: '(forced-colors: none active)' 
PASS Should not be parseable in JS: '(forced-colors: active/none)' 
FAIL Check that none evaluates to false in the boolean context assert_equals: expected true but got false
